Move Metal related files to MetalModule
