+++ Modules/MetalModule/Tests/MetalModuleTests/MetalModuleTests.swift
@@ -1,0 +1,8 @@
+import Testing
+@testable import MetalModule
+
+@Test func example() async throws {
+    // Write your test here and use APIs like `#expect(...)` to check expected conditions.
+    // Swift Testing Documentation
+    // https://developer.apple.com/documentation/testing
+}
