+++ OptiUniverse.xcworkspace/contents.xcworkspacedata
@@ -4,4 +4,11 @@
    <FileRef
       location = "container:OptiUniverse.xcodeproj">
    </FileRef>
+   <Group
+      location = "group:Modules"
+      name = "Modules">
+      <FileRef
+         location = "group:MetalModule">
+      </FileRef>
+   </Group>
 </Workspace>
