--- OptiUniverse/Features/Metal/Assets/Models/high_resolution_solar_system.usdz
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:ce2497b4ce1d6564999a4d56956cd3383e53317e1e483ceda3ee04d5084d47ef
-size 58826158
